Fixed `huddle::call`.
The way how `huddle::call` used its `arguments` argument did not
work in the sense that wrong arguments get dispatched through
the callback ensemble (at least in Tcl 8.6). The `arguments`
was changed to `args` and extra `eval` added to expand the
variable length argument list back into individual arguments.

Tests were added to exercise some basic functionality of `huddle::call`.

It remains a question if this routine is anyhow useful as users need
to know the exact prototype of the targeted, type-specific command.
If you take `huddle::types::dict` as an example, some of its routines
take `huddle_object` while others take `src` (which is supposed to
be the result of the `huddle::get_src <huddle_obj>` call).

It would make more sense to require a common convention that all
type specific routines take either a huddle object or a huddle
node as their first argument. Doing this would make the use of
`huddle::call` more convenient and predictable.